Creating scaffolding and using lottie
