+++ client/src/components/Header.module.css
@@ -1,22 +1,21 @@
-body{
-    margin: 0px;
+body {
+  margin: 0px;
 }
 .content {
-    height: 38px;
-    background-color: #ffd60a;
-    display: flex;
-    justify-content: end;
-    border: 0.8px groove #c2c2c24f;
-    background-color: #ffd60a;
-
-
+  height: 38px;
+  background-color: #ffd60a;
+  display: flex;
+  justify-content: end;
+  border: 0.8px groove #c2c2c24f;
+  background-color: #ffd60a;
 }
 .nav {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    margin-right: 10px;
-    font-size: 10px;
-    font-weight: 600;
-    color: black;
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  margin-right: 10px;
+  font-size: 10px;
+  font-weight: 600;
+  color: black;
 }
+/* @media (max-width: ); */
